refactor to use one component
