+++ app/src/features/observations/BlockObservationPage.tsx
@@ -8,7 +8,10 @@ import { makeStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import { Formik, FormikProps } from 'formik';
 import React, { useRef, useState, useContext, useCallback, useEffect } from 'react';
-import BlockObservationForm, { BlockObservationInitialValues } from './components/BlockObservationForm';
+import BlockObservationForm, {
+  BlockObservationInitialValues,
+  IBlockObservationForm
+} from './components/BlockObservationForm';
 import { Prompt, useHistory, useParams } from 'react-router';
 import { DialogContext } from 'contexts/dialogContext';
 import { AddBlockObservationI18N } from 'constants/i18n';
@@ -47,15 +50,19 @@ const BlockObservationPage = () => {
   // Ability to bypass showing the 'Are you sure you want to cancel' dialog
   const [enableCancelCheck] = useState(true);
   const [tableData] = useState<any[][]>([[, , , , , , , , , , , , , ,]]);
-  const [initialValues] = useState(BlockObservationInitialValues);
 
   const [isLoadingProject, setIsLoadingProject] = useState(true);
   const [isLoadingSurvey, setIsLoadingSurvey] = useState(true);
+  const [isLoadingObservation, setIsLoadingObservation] = useState(true);
   const [projectWithDetails, setProjectWithDetails] = useState<IGetProjectForViewResponse | null>(null);
   const [surveyWithDetails, setSurveyWithDetails] = useState<IGetSurveyForViewResponse | null>(null);
+  const [observationWithDetails, setObservationWithDetails] = useState<IBlockObservationForm>(
+    BlockObservationInitialValues
+  );
 
   const projectId = urlParams['id'];
   const surveyId = urlParams['survey_id'];
+  const observationId = urlParams['observation_id'];
 
   const getProject = useCallback(async () => {
     const projectWithDetailsResponse = await biohubApi.project.getProjectForView(projectId);
@@ -76,6 +83,21 @@ const BlockObservationPage = () => {
     setSurveyWithDetails(surveyWithDetailsResponse);
   }, [biohubApi.survey, urlParams]);
 
+  const getObservation = useCallback(async () => {
+    const observationWithDetailsResponse = await biohubApi.observation.getObservationForUpdate(
+      projectId,
+      surveyId,
+      observationId,
+      'block'
+    );
+
+    if (!observationWithDetailsResponse || !observationWithDetailsResponse.data) {
+      return;
+    }
+
+    setObservationWithDetails(observationWithDetailsResponse.data.metaData);
+  }, [biohubApi.observation, urlParams]);
+
   useEffect(() => {
     if (isLoadingProject && !projectWithDetails) {
       getProject();
@@ -89,6 +111,13 @@ const BlockObservationPage = () => {
       setIsLoadingSurvey(false);
     }
   }, [isLoadingSurvey, surveyWithDetails, getSurvey]);
+
+  useEffect(() => {
+    if (isLoadingObservation && observationId) {
+      getObservation();
+      setIsLoadingObservation(false);
+    }
+  }, [observationId, isLoadingObservation, observationWithDetails, getObservation]);
 
   const defaultCancelDialogProps = {
     dialogTitle: AddBlockObservationI18N.cancelTitle,
@@ -137,7 +166,7 @@ const BlockObservationPage = () => {
     return true;
   };
 
-  if (!projectWithDetails || !surveyWithDetails) {
+  if (!projectWithDetails || !surveyWithDetails || !observationWithDetails) {
     return <CircularProgress className="pageProgress" size={40} />;
   }
 
@@ -169,12 +198,12 @@ const BlockObservationPage = () => {
                 className={classes.breadCrumbLink}>
                 <Typography variant="body2">{surveyWithDetails.survey_details.survey_name}</Typography>
               </Link>
-              <Typography variant="body2">Add Block Observation</Typography>
+              <Typography variant="body2">{observationId ? 'Edit' : 'Add'} Block Observation</Typography>
             </Breadcrumbs>
           </Box>
           <Box mb={3}>
             <Typography data-testid="block-observation-heading" variant="h1">
-              Add Block Observation
+              {observationId ? 'Edit' : 'Add'} Block Observation
             </Typography>
           </Box>
           <Box mb={5}>
@@ -186,7 +215,7 @@ const BlockObservationPage = () => {
           <Box pl={3} pr={3} component={Paper} display="block">
             <Formik
               innerRef={formikRef}
-              initialValues={initialValues}
+              initialValues={observationWithDetails}
               enableReinitialize={true}
               validateOnBlur={false}
               validateOnChange={false}
@@ -200,16 +229,18 @@ const BlockObservationPage = () => {
                 color="primary"
                 onClick={() => {}}
                 className={classes.actionButton}>
-                Save and Exit
+                Save {observationId ? 'Changes' : 'and Exit'}
               </Button>
-              <Button
-                type="submit"
-                variant="contained"
-                color="primary"
-                onClick={() => {}}
-                className={classes.actionButton}>
-                Save and Next Block
-              </Button>
+              {!observationId && (
+                <Button
+                  type="submit"
+                  variant="contained"
+                  color="primary"
+                  onClick={() => {}}
+                  className={classes.actionButton}>
+                  Save and Next Block
+                </Button>
+              )}
               <Button variant="outlined" color="primary" onClick={handleCancel} className={classes.actionButton}>
                 Cancel
               </Button>
